Move views and forms into separate app
